requirements.txt
anyio==3.6.1
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
fastapi==0.79.0
h11==0.13.0
idna==3.3
install==1.3.5
psycopg2==2.9.3
pydantic==1.9.2
PyJWT==2.4.0
requests==2.28.1
sniffio==1.2.0
starlette==0.19.1
typing_extensions==4.3.0
urllib3==1.26.11
uvicorn==0.18.2
websockets==10.3
